requirements.txt
socketIO-client==0.6.5
pyserial==2.7